mod duo driver issue
my mod duo has had trouble connecting to any of the three windows pcs i'm using. the first day it didn't connect until i restarted it a few times and i was able to update the pedal. i made a few boards and since then, the pedal cannot be accessed through any computer. is there a way to reset the device and i can attempt to reinstall the update? it also doesn't actually go into the system update mode most of the time; usually it just freezes and never is ready to connect to the pc for copying the update.